Add file upload, expiry timer, and sharing link feature

+++ src/main/java/com/company/fileSharingManagement/service/FileService.java
@@ -11,10 +11,16 @@ import com.company.fileSharingManagement.model.FileModel;
 
 
 public interface FileService {
-        public ResponseEntity<?> uploadFile(MultipartFile file, String uploadedBy) throws IOException;
-        public ResponseEntity<?> getFile(int id);
-        public ResponseEntity<?> deleteFile(int id) ;
-        public void deleteExpiredFiles();
-       public List<FileModel> getAllFiles();
-        public ResponseEntity<?> shareFile(int id);
+
+	public ResponseEntity<?> uploadFile(MultipartFile file, String uploadedBy) throws IOException;
+
+	public ResponseEntity<?> getFile(Long id);  
+
+	public ResponseEntity<?> deleteFile(Long id);  
+
+	public void deleteExpiredFiles();
+
+	public List<FileModel> getAllFiles();
+
+	public ResponseEntity<?> shareFile(Long id);  
 }
